requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
numpy
joblib
scikit-learn
